pyproject.toml
[project]
name = "coqueiral-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Jean Lucas",email = "jeanlucas2504@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.0)",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'poetry run ruff check . && ruff check . --diff'
format = 'poetry run ruff check . --fix && ruff format .'
run = 'poetry run fastapi dev coqueiral_fastapi/auth_service/app/main.py'
pre_test = 'poetry run task lint'
test = 'poetry run pytest -s -x --cov=coqueiral_fastapi -vv'
post_test = 'poetry run coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
